Refactored these tests to be unit tests, because we should encourage encapulation. Having these tests in integration tests forced me to make things public that shouldn't have been.
