+++ tests/common/mod.rs
@@ -1,8 +1,2 @@
-pub mod collections;
-pub use collections::*;
-
-pub mod model_generators;
-pub use model_generators::*;
-
 pub mod error;
 pub use error::*;
